field의 preset을 기준으로 검색의 input type select분기 처리 추가

+++ bluesky-parent/bluesky-web/bluesky-web-dynamiccrud/src/main/java/net/luversof/web/dynamiccrud/setting/jdbc/mapper/mariadb/SubMenuRowMapper.java
@@ -15,7 +15,7 @@ public class SubMenuRowMapper extends SettingRowMapper<SubMenu> {
 			rs.getString("subMenu"),
 			rs.getString("subMenuName"),
 			rs.getString("template"),
-			rs.getInt("displayOrder"),
+			rs.getShort("displayOrder"),
 			rs.getShort("groupNo"),
 			rs.getString("groupTemplate"),
 			rs.getShort("pageSize"),
